refactor(richmond-buildings): split ArcFloor into behavior submodules
Break the monolithic `arc_floor.rs` into modern Rust submodules
(`ring`, `walls`, `openings`, `slabs`, `tests`) with Layer 1/2 and
mapping logic as `impl ArcFloor` / `impl ArcFloorParams`, keeping the
parent file for types, builder, and construct orchestration.

## Agent Dialogue

- **Developer:** Split `arc_floor.rs` into submodules with behavior `impl`s (e.g. openings); no `mod.rs`.
- **Agent:** Added `arc_floor/{ring,walls,openings,slabs,tests}.rs`; public API unchanged.
- **Developer:** Write the commit file.
- **Agent:** Wrote this commit-ready summary for HEAD `71d71659`.
